Make imgcat.mpl_backend compatible with matplotlib 3.6+ (#13)
Matplotlib 3.6+ [1] requires FigureCanvas of a matplotlib backend module
to have `required_interactive_framework`, and we had an warning message
because the module was not exporting FigureCanvas (#13).

I have no idea why I had to comment this in the past (see 99aaedc),
but probably there was some bug with older matplotlib versions. This is
no longer reproducible so I assume imgcat would work OK and reasonably
well with recent versions of matplotlib.

[1]: https://matplotlib.org/stable/api/prev_api_changes/api_changes_3.6.0.html

Fixes #13.